# Speed up `fmpz_mul_si`