--- a/fmpz/mul_si.c
+++ b/fmpz/mul_si.c
@@ -19,27 +19,24 @@ fmpz_mul_si(fmpz_t f, const fmpz_t g, slong x)
 {
     fmpz c2 = *g;
 
-    if (x == 0)
-    {
-        fmpz_zero(f);
-        return;
-    }
-    else if (!COEFF_IS_MPZ(c2)) /* c2 is small */
+    if (!COEFF_IS_MPZ(c2)) /* c2 is small */
     {
         mp_limb_t prod[2];
-        mp_limb_t uc2 = FLINT_ABS(c2);
-        mp_limb_t ux = FLINT_ABS(x);
+        mp_limb_t mc2 = c2;
+        mp_limb_t mx = x;
 
-        /* unsigned limb by limb multiply (assembly for most CPU's) */
-        umul_ppmm(prod[1], prod[0], uc2, ux);
-        if ((c2 ^ x) < WORD(0))
-            fmpz_neg_uiui(f, prod[1], prod[0]);
-        else
-            fmpz_set_uiui(f, prod[1], prod[0]);
+        /* limb by limb multiply (assembly for most CPU's) */
+        smul_ppmm(prod[1], prod[0], mc2, mx);
+        fmpz_set_signed_uiui(f, prod[1], prod[0]);
     }
     else                        /* c2 is large */
     {
-        __mpz_struct *mpz_ptr = _fmpz_promote(f);   /* ok without val as if aliased both are large */
-        flint_mpz_mul_si(mpz_ptr, COEFF_TO_PTR(c2), x);
+        if (x == 0)
+            fmpz_zero(f);
+        else
+        {
+            __mpz_struct *mpz_ptr = _fmpz_promote(f);   /* ok without val as if aliased both are large */
+            flint_mpz_mul_si(mpz_ptr, COEFF_TO_PTR(c2), x);
+        }
     }
 }