By running `fmpz/profile/p-mul_si.c`, I got
```
fmpz_mul_si_new min time is 8.770 cycles, max time is 9.621 cycles
fmpz_mul_si_old min time is 13.057 cycles, max time is 15.946 cycles
```
which is a nice speedup.

Let me know if the inlining is not a good idea or whatever, or if perhaps the test results are misleading.

Edit: This was on a Haswell 4770k. Let me know if you get something that does not represent this result.