@albinahlback 2021-11-05
I think very similar speed ups can be made for the other `fmpz` multiplications.