@wbhart 2021-11-16
Can this be done without moving the check for x == 0? This is a sufficiently important case that it should be optimised for.