@wbhart 2021-11-16
@tthsqe12 Can you say why it would be good to slow down one of the most important functions in the case it is going to be used the most?