@albinahlback 2021-11-16
> Can this be done without moving the check for x == 0? This is a sufficiently important case that it should be optimised for.

Yeah sure, but it makes it faster for when the `fmpz` is an `mpz`. Otherwise this is actually faster.